# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse:
    properties:
      error:
        example: Internal Server Error
        type: string
    type: object
  github_com_scythe504_rbac-finance-processing_internal_utils.MessageResponse:
    properties:
      message:
        example: success
        type: string
    type: object
  internal_server.AuthResponse:
    properties:
      token:
        example: jwt.token.here
        type: string
      user_id:
        example: "1234"
        type: string
    type: object
...
  internal_server.LoginUser:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  internal_server.RegisterUser:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
...
  internal_server.UserDetailsResponse:
    properties:
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        type: string
      role:
        example: admin
        type: string
      user_id:
        example: "1234"
        type: string
    type: object
info:
  contact: {}
  description: API for managing personal finance with Role-Based Access Control.
  title: RBAC Finance Processing API
  version: 1.0.0
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      operationId: loginUser
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_server.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/internal_server.AuthResponse'
        "400":
          description: Invalid Request Body
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "401":
          description: Invalid Credentials
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      summary: Login a user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new account with a specific role
      operationId: registerUser
      parameters:
      - description: User registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_server.RegisterUser'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/internal_server.AuthResponse'
        "400":
          description: Invalid Request Body / Email Address
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "401":
          description: Self registration is only available for viewer role. Contact
            an admin for desired role.
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "409":
          description: User with email already exists, Please Login
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
...
  /me:
    get:
      consumes:
      - application/json
      description: Get current logged in user information
      operationId: getUserDetails
      produces:
      - application/json
      responses:
        "200":
          description: User details retrieved
          schema:
            $ref: '#/definitions/internal_server.UserDetailsResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user details
      tags:
      - Users
...
  /users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Update the role of a specific user (Admin only)
      operationId: setUserRole
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Target role details
        in: body
        name: body
        required: true
        schema:
          properties:
            target_user_role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.MessageResponse'
        "400":
          description: Invalid Request Body / Role
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set user role
      tags:
      - Users
  /users/{id}/status:
    patch:
      consumes:
      - application/json
      description: Enable or disable a user account (Admin only)
      operationId: toggleUserStatus
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle user status
      tags:
      - Users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"